move screens client services back in FrontendServiceClient
