+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/frontend/FrontendServiceClient.java
@@ -93,6 +93,17 @@ public class FrontendServiceClient extends VPinStudioClientService {
     return getRestClient().get(API + API_SEGMENT_FRONTEND + "/media/" + gameId, FrontendMediaRepresentation.class);
   }
 
+  public FrontendPlayerDisplay getScreenDisplay(VPinScreen screen) {
+    return getRestClient().get(API + API_SEGMENT_FRONTEND + "/screen/" + screen.name(), FrontendPlayerDisplay.class);
+  }
+/*
+  OLE: Not more used at the moment  
+  public List<FrontendPlayerDisplay> getScreenDisplays() {
+    FrontendPlayerDisplay[] displays = getRestClient().get(API + API_SEGMENT_FRONTEND + "/screens", FrontendPlayerDisplay[].class);
+    return displays != null ? Arrays.asList(displays) : Collections.emptyList();
+  }
+  */
+
   public GameEmulatorRepresentation getDefaultGameEmulator() {
     List<GameEmulatorRepresentation> gameEmulators = getGameEmulators();
     return gameEmulators.size() > 0 ? gameEmulators.get(0) : null;
